pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "frrsd"
version = "0.0.0"
authors = [
  { name="David Timber", email="dxdt@dev.snart.me" },
]
description = "Collect and plot BGP summary and ping data locally on the router"
readme = "README.md"
requires-python = ">= 3.11"
dependencies = [ "psycopg2-binary", "pyjson5", "matplotlib" ]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/dxdxdt/frrsd/"
Issues = "https://github.com/dxdxdt/frrsd/issues"
